+++ packages/bluesky/lib/firehose.dart
@@ -4,7 +4,4 @@
 
 export 'package:atproto/firehose.dart';
 
-export 'package:bluesky/src/tools/utils/repo_commit_handler.dart';
-export 'package:bluesky/src/tools/utils/repo_commit_create.dart';
-export 'package:bluesky/src/tools/utils/repo_commit_delete.dart';
-export 'package:bluesky/src/tools/utils/repo_commit_update.dart';
+export 'package:bluesky/src/services/codegen/repo_commit_handler.dart';